requirements.txt
astroid==1.6.3
backports.functools-lru-cache==1.5
configparser==3.5.0
enum34==1.1.6
futures==3.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.14.2
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
Pillow==5.1.0
pylint==1.8.4
singledispatch==3.4.0.3
six==1.11.0
wrapt==1.10.11
